requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
gunicorn==20.1.0
idna==3.4
Pillow==9.2.0
python-dotenv==0.21.0
python-stripe==0.0.5
requests==2.28.1
sqlparse==0.4.3
stripe==4.2.0
urllib3==1.26.12
python-stripe
